+++ lib/models/event.dart
@@ -1,5 +1,5 @@
-import 'package:fomo/widgets/event_cell.dart';
-import 'package:fomo/widgets/mini_event_Cell.dart';
+import 'package:fomo/widgets/Events/mini_event_cell.dart';
+import 'package:fomo/widgets/widgets.dart';
 
 class Event {
   final String name;
