+++ lib/widgets/Common/vertical_list.dart
@@ -1,4 +1,4 @@
-import 'widgets.dart';
+import '../widgets.dart';
 import 'package:flutter/material.dart';
 
 class VerticalList extends StatelessWidget {
